+++ src/main/java/com/SolucionesInformaticasBA/minimarket/modules/ventas/api/dto/ResumenDiarioResponse.java
@@ -1,5 +1,7 @@
 package com.SolucionesInformaticasBA.minimarket.modules.ventas.api.dto;
 
+import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 import lombok.Builder;
@@ -10,8 +12,8 @@ import lombok.Data;
 public class ResumenDiarioResponse {
     private LocalDate fecha;
     private int cantidadVentas;
-    private float totalVentas;
-    private float totalEfectivo;
-    private float totalTarjeta;
-    private float totalTransferencia;
+    private BigDecimal totalVentas;
+    private BigDecimal totalEfectivo;
+    private BigDecimal totalTarjeta;
+    private BigDecimal totalTransferencia;
 }
